requirements.txt
python-socketio[client]>=5.0.0